feat: release 0.5.3 (#16)
* chore: update deps

* docs: update docs

* chore(ci): bumps nightly versions in ci toolchain

+++ src/socket/tcp.rs
@@ -141,10 +141,10 @@ impl TcpSocket {
                 self.recv_flags.as_i32(),
             )
         };
-        if (result as i32) < 0 {
+        if result < 0 {
             Err(SocketError::Errno(unsafe { sys::sceNetInetGetErrno() }))
         } else {
-            Ok(result)
+            Ok(result as usize)
         }
     }
 
@@ -254,6 +254,13 @@ impl<'a> Open<'a> for TcpSocket {
 
 impl Read for TcpSocket {
     /// Read from the socket
+    ///
+    /// # Parameters
+    /// - `buf`: The buffer where the read data will be stored
+    ///
+    /// # Returns
+    /// - `Ok(usize)` if the read was successful. The number of bytes read
+    /// - `Err(SocketError)` if the read was unsuccessful.
     ///
     /// # Errors
     /// - [`SocketError::NotConnected`] if the socket is not connected
